delete old socket implementation
